sample dto now created/updated and sent
- changed all fields within the api with the type number to doubles on Dto model
- added sample extenstion to both update and create sample with location dto

+++ src/ImuExports/Tasks/AusGeochem/Models/Api/LocationDto.cs
@@ -29,7 +29,7 @@ public class LocationDto
         public double? Lat { get; set; }
 
         [JsonPropertyName("latLonPrecision")]
-        public int? LatLonPrecision { get; set; }
+        public double? LatLonPrecision { get; set; }
 
         [JsonPropertyName("lon")]
         public double? Lon { get; set; }
